% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{myShinyApp2}
\alias{myShinyApp2}
\title{simple function 2}
\usage{
myShinyApp2()
}
\value{
dummy
}
\description{
it tries to load and unload \code{shiny} before and after the execution of the module.
You can disable the same by setting \code{options(AutoUnloadShiny = FALSE)}.
}
