% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{myShinyApp3}
\alias{myShinyApp3}
\title{simple function 3}
\usage{
myShinyApp3()
}
\value{
dummy
}
\description{
it tries to load and unload \code{shiny} before and after the execution of the module.
You can disable the same by setting \code{options(AutoUnloadShiny = FALSE)}.
}
